[MSA-28] feat: HubDataInitializer 와 HubConnectionDataInitializer 하나의 클래스로
