--- hub/src/main/java/com/i4u/hub/infrastructure/config/HubConnectionDataInitializer.java
@@ -1,115 +1,0 @@
-package com.i4u.hub.infrastructure.config;
-
-import com.i4u.hub.domain.model.Hub;
-import com.i4u.hub.domain.model.HubConnection;
-import com.i4u.hub.domain.repository.HubConnectionRepository;
-import com.i4u.hub.domain.repository.HubRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.DependsOn;
-import org.springframework.core.annotation.Order;
-
-import java.util.*;
-import java.util.stream.Collectors;
-
-@Slf4j
-@Configuration
-@RequiredArgsConstructor
-@Order(2) // Hub 데이터가 먼저 저장된 후 실행되도록 순서 지정
-public class HubConnectionDataInitializer {
-
-    private final HubRepository hubRepository;
-    private final HubConnectionRepository hubConnectionRepository;
-
-    @Bean
-    @DependsOn("initHubData")
-    public CommandLineRunner initHubConnectionData() {
-        return args -> {
-            try {
-                log.info("Hub Connection 데이터 초기화 시작");
-
-                // 기존 데이터 확인
-                long count = hubConnectionRepository.count();
-                if (count > 0) {
-                    log.info("이미 {} 개의 Hub Connection 데이터가 존재합니다. 초기화를 건너뜁니다.", count);
-                    return;
-                }
-
-                List<Hub> allHubs = hubRepository.findAll();
-                if (allHubs.isEmpty()) {
-                    log.warn("Hub 데이터가 존재하지 않습니다. Hub Connection 초기화를 중단합니다.");
-                    return;
-                }
-
-                log.info("총 {} 개의 Hub 데이터를 읽었습니다.", allHubs.size());
-
-                Map<String, Hub> hubMap = allHubs.stream()
-                        .collect(Collectors.toMap(Hub::getHubName, hub -> hub));
-
-                List<HubConnection> connections = new ArrayList<>();
-
-                // 경기 남부 센터 연결
-                addConnection(connections, hubMap, "경기 남부 센터", "경기 북부 센터", 90, 69);
-                addConnection(connections, hubMap, "경기 남부 센터", "서울특별시 센터", 52, 52);
-                addConnection(connections, hubMap, "경기 남부 센터", "인천광역시 센터", 80, 66);
-                addConnection(connections, hubMap, "경기 남부 센터", "강원특별자치도 센터", 139, 119);
-                addConnection(connections, hubMap, "경기 남부 센터", "경상북도 센터", 186, 155);
-                addConnection(connections, hubMap, "경기 남부 센터", "대전광역시 센터", 112, 96);
-                addConnection(connections, hubMap, "경기 남부 센터", "대구광역시 센터", 240, 187);
-
-                // 대전광역시 센터 연결
-                addConnection(connections, hubMap, "대전광역시 센터", "충청남도 센터", 87, 63);
-                addConnection(connections, hubMap, "대전광역시 센터", "충청북도 센터", 39, 46);
-                addConnection(connections, hubMap, "대전광역시 센터", "세종특별자치시 센터", 22, 25);
-                addConnection(connections, hubMap, "대전광역시 센터", "전북특별자치도 센터", 85, 74);
-                addConnection(connections, hubMap, "대전광역시 센터", "광주광역시 센터", 168, 113);
-                addConnection(connections, hubMap, "대전광역시 센터", "전라남도 센터", 229, 152);
-                addConnection(connections, hubMap, "대전광역시 센터", "경기 남부 센터", 110, 86);
-                addConnection(connections, hubMap, "대전광역시 센터", "대구광역시 센터", 151, 106);
-
-                // 대구광역시 센터 연결
-                addConnection(connections, hubMap, "대구광역시 센터", "경상북도 센터", 112, 74);
-                addConnection(connections, hubMap, "대구광역시 센터", "경상남도 센터", 94, 75);
-                addConnection(connections, hubMap, "대구광역시 센터", "부산광역시 센터", 112, 81);
-                addConnection(connections, hubMap, "대구광역시 센터", "울산광역시 센터", 109, 87);
-                addConnection(connections, hubMap, "대구광역시 센터", "경기 남부 센터", 235, 155);
-                addConnection(connections, hubMap, "대구광역시 센터", "대전광역시 센터", 152, 105);
-
-                // 경상북도 센터 연결
-                addConnection(connections, hubMap, "경상북도 센터", "경기 남부 센터", 200, 132);
-                addConnection(connections, hubMap, "경상북도 센터", "대구광역시 센터", 104, 75);
-
-                log.info("총 {} 개의 Hub Connection 데이터를 저장합니다.", connections.size());
-                hubConnectionRepository.saveAll(connections);
-                log.info("Hub Connection 데이터 초기화 완료");
-            } catch (Exception e) {
-                log.error("Hub Connection 데이터 초기화 중 오류 발생: {}", e.getMessage(), e);
-                throw e;
-            }
-        };
-    }
-
-    private void addConnection(List<HubConnection> connections, Map<String, Hub> hubMap,
-                               String departureName, String arrivalName, Integer distance, Integer timeMinutes) {
-        Hub departure = hubMap.get(departureName);
-        Hub arrival = hubMap.get(arrivalName);
-
-        if (departure != null && arrival != null) {
-            HubConnection connection = HubConnection.builder()
-                    .departureHub(departure)
-                    .arrivalHub(arrival)
-                    .distance(distance)
-                    .hubToHubTime(timeMinutes)
-                    .build();
-
-            connections.add(connection);
-            log.debug("연결 추가: {} → {}, 거리: {}km, 시간: {}분",
-                    departureName, arrivalName, distance, timeMinutes);
-        } else {
-            log.warn("연결 추가 실패: {} → {}, 허브가 존재하지 않습니다.", departureName, arrivalName);
-        }
-    }
-}
